+++ app/javascript/packages/components/index.js
@@ -1,4 +1,5 @@
 export { default as Alert } from './alert';
+export { default as Button } from './button';
 export { default as BlockLink } from './block-link';
 export { default as Icon } from './icon';
 export { default as SpinnerDots } from './spinner-dots';
